Barchart component created
