+++ src/App.js
@@ -1,21 +1,12 @@
 import React from "react";
 import "./style.less";
 
+// CSS Import for React-Vis
 import "react-vis/dist/style.css";
-
-import {
-  XYPlot,
-  XAxis,
-  YAxis,
-  VerticalGridLines,
-  HorizontalGridLines,
-  VerticalBarSeries,
-  VerticalBarSeriesCanvas,
-  LabelSeries
-} from "react-vis";
 
 // Component Imports
 import Navbar from "./components/navbar/Navbar.js";
+import Barchart from "./components/barchart/Barchart.js";
 
 class App extends React.Component {
   constructor(props) {
@@ -26,30 +17,11 @@ class App extends React.Component {
   }
 
   render() {
-    const greenData = [{ x: "A", y: 10 }, { x: "B", y: 5 }, { x: "C", y: 15 }];
-
-    const blueData = [{ x: "A", y: 12 }, { x: "B", y: 2 }, { x: "C", y: 11 }];
-
-    const labelData = greenData.map((d, idx) => ({
-      x: d.x,
-      y: Math.max(greenData[idx].y, blueData[idx].y)
-    }));
-
-    const { useCanvas } = this.state;
-    const content = useCanvas ? "TOGGLE TO SVG" : "TOGGLE TO CANVAS";
-    const BarSeries = useCanvas ? VerticalBarSeriesCanvas : VerticalBarSeries;
-
     return (
       <div>
         <Navbar />
         <h1>Mayaskara</h1>
-        <XYPlot xType="ordinal" width={300} height={300} xDistance={100}>
-          <XAxis />
-          <YAxis />
-          <BarSeries className="vertical-bar-series-example" data={greenData} />
-          <BarSeries data={blueData} />
-          <LabelSeries data={labelData} getLabel={d => d.x} />
-        </XYPlot>
+        <Barchart />
       </div>
     );
   }
